Misc. minor cleanup and merge
